Fix prefix path url (for previous commit)
Previous commit failed to fix the problem, this one fixed it and tested
on production site (this is the reason previous commit failed as it is
not tested on production site).

Turn out we need prefix. It needs to be aligned with url that has been
set on dashboard. It's separated from what has been set on server side
(proxy & reverse proxy).