Generate translated text to file. Process text.
Much of text processings usually via regex to process text to show on
generated html file are added in this changes.

Dynamically find a first image that found in the article and use it as
an article thumbnail.

html-to-text is fetched from forked repo on github, not npm. See
package.json. This is due to we need to develop it in parallel.

Update .gitignore to reflect the need to watch node_modules in local
machine to reduce time needed to debug html-to-text as with reason above.